requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
tabulate==0.9.0
python-docx==1.1.2
